pyproject.toml
[project]
name = "wikitalkie"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-adk>=1.20.0",
    "litellm>=1.80.9",
]
